requirements.txt
aiohttp
aiosqlite
discord.py
tabulate
gspread

dpytest